services/guard/requirements.txt
fastapi
uvicorn[standard]
kafka-python
python-dotenv
psycopg2-binary
prometheus-client
